refactors so each card is separate form
